clean code

+++ src/main/java/com/iop/productservice/controller/ProductController.java
@@ -38,7 +38,7 @@ public class ProductController {
      * ProductNotFoundException
      */
     @GetMapping("/{productId}")
-    public ResponseEntity<Product> getProduct(@PathVariable Integer productId) {
+    public ResponseEntity<Product> getProduct(@PathVariable Long productId) {
         Product product = productService.getProduct(productId);
         if (product == null) {
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
@@ -67,7 +67,7 @@ public class ProductController {
      * @return updated product information along with HTTP Status code
      */
     @PatchMapping("/{productId}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Integer productId,
+    public ResponseEntity<Product> updateProduct(@PathVariable Long productId,
                                                  @RequestBody ProductUpdateRequest updateRequest)
     {
         return productService.updateProduct(productId, updateRequest);
@@ -82,7 +82,7 @@ public class ProductController {
      */
     @DeleteMapping("/{productId}")
     public ResponseEntity<ResponseMessage> deleteProduct(
-            @PathVariable Integer productId) {
+            @PathVariable Long productId) {
         return productService.deleteProduct(productId);
     }
