+++ src/main/java/com/iop/productservice/repository/ProductRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
     // Named query
-    Optional<Product> findByProductIdAndIsDeleted(Integer id, boolean isDeleted);
+    Optional<Product> findByProductIdAndIsDeleted(Long id, boolean isDeleted);
 }
